Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-jose[cryptography]
